+++ eval_anything/benchmarks/text_to_text/HumanEval/eval.py
@@ -8,7 +8,8 @@ from eval_anything.models.base_model import BaseModel
 from eval_anything.utils.logger import EvalLogger
 from eval_anything.utils.register import BenchmarkRegistry
 from eval_anything.utils.cache_manager import CacheManager
-
+from eval_anything.utils.data_type import InferenceInput
+from eval_anything.benchmarks.text_to_text.HumanEval.metrics import PassAtK
 @BenchmarkRegistry.register('HumanEval')
 class HumanEvalBenchmark(T2TBenchmark):
     def __init__(self, 
@@ -23,39 +24,13 @@ class HumanEvalBenchmark(T2TBenchmark):
         self.benchmark_name = "HumanEval"
         self.benchmark_cfgs = self.get_benchmark_cfgs(self.benchmark_name)
 
-    def run(self, 
-            task_list: list[str],) -> tuple[dict[str, list[EvaluationResult]], dict[str, dict[str, float]], dict[str, dict[str, float]]]:
-        """Run benchmark
-        Args:
-            task_list (list[str]): task list
-            
-        Returns:
-            evaluation_details (dict[str, list[EvaluationResult]]): evaluation details
-            evaluation_results (dict[str, dict[str, float]]): evaluation results
-        """
-        self.logger.log('info', f'Evaluating {self.benchmark_name}...')
+    def to_InferenceInput(self, task_list: list[str]) -> dict[str, list[InferenceInput]]:
+        """Convert a task list to a InferenceInput dict instances"""
+        input_data = super().to_InferenceInput(task_list)
 
-        dataloader = self.init_dataloader(self.eval_cfgs, self.benchmark_cfgs)
-        input_data = dataloader.load_dataset(task_list)    # Input_data: list[InferenceInput]
-        metadata = []
-        for key, value in input_data.items():
-            metadata = [item.metadata for item in value]
+        for task, inference_input in input_data.items():
+            metadata = [item.metadata for item in inference_input]
+        for inference_input_item, metadata_item in zip(inference_input, metadata):
+            inference_input_item.ref_answer = metadata_item
         
-        inference_outputs = self.batch_inference(self.model, input_data)
-        
-        evaluation_details = {}
-        evaluation_results = {}
-
-        for task in task_list:
-            evaluation_details[task], evaluation_results[task] = self.calculate_metrics(self.benchmark_name, inference_outputs[task], metadata, self.benchmark_cfgs.answer_extractor, self.benchmark_cfgs.metrics)
-
-        if len(task_list) > 1:
-            overall_result = self.calculate_overall_metrics(self.benchmark_cfgs.overall_metrics, result=evaluation_results)
-        else:
-            overall_result = {}
-            
-        self.display_benchmark_results(self.benchmark_name, evaluation_results)
-        if overall_result != {}:
-            self.display_benchmark_results(self.benchmark_name, overall_result)
-        self.save_benchmark_details(self.output_path, self.benchmark_name, input_data, evaluation_details)
-        return evaluation_details, evaluation_results, overall_result
+        return input_data
