+++ eval_anything/benchmarks/text_to_text/gsm8k/eval.py
@@ -3,12 +3,12 @@
 """
 from eval_anything.pipeline.t2t_benchmark import T2TBenchmark
 from collections import namedtuple
-from eval_anything.utils.data_type import EvaluationResult
 from eval_anything.evaluate_tools.t2t_tools import RegexMatchNumber
 from eval_anything.models.base_model import BaseModel
 from eval_anything.utils.logger import EvalLogger
 from eval_anything.utils.register import BenchmarkRegistry
 from eval_anything.utils.cache_manager import CacheManager
+from eval_anything.utils.data_type import InferenceInput
 
 @BenchmarkRegistry.register('gsm8k')
 class gsm8kBenchmark(T2TBenchmark):
@@ -24,37 +24,12 @@ class gsm8kBenchmark(T2TBenchmark):
         self.benchmark_name = "gsm8k"
         self.benchmark_cfgs = self.get_benchmark_cfgs(self.benchmark_name)
 
-    def run(self, 
-            task_list: list[str],) -> tuple[dict[str, list[EvaluationResult]], dict[str, dict[str, float]], dict[str, dict[str, float]]]:
-        """Run benchmark
-        Args:
-            task_list (list[str]): task list
-            
-        Returns:
-            evaluation_details (dict[str, list[EvaluationResult]]): evaluation details
-            evaluation_results (dict[str, dict[str, float]]): evaluation results
-        """
-        self.logger.log('info', f'Evaluating {self.benchmark_name}...')
-
-        dataloader = self.init_dataloader(self.eval_cfgs, self.benchmark_cfgs)
-        input_data = dataloader.load_dataset(task_list)    # Input_data: list[InferenceInput]
+    def to_InferenceInput(self, task_list: list[str]) -> dict[str, list[InferenceInput]]:
+        """Convert a task list to a list of InferenceInput dict instances"""
+        input_data = super().to_InferenceInput(task_list)
         
-        inference_outputs = self.batch_inference(self.model, input_data)
-        ref_answers = {task: self.get_ref_answer(input_data[task], inference_outputs[task]) for task in task_list}
-        evaluation_details = {}
-        evaluation_results = {}
-
-        for task in task_list:
-            ground_truth = RegexMatchNumber(additional_pattern=r"{original_pattern}", match_index=-1).apply(ref_answers[task])
-            evaluation_details[task], evaluation_results[task] = self.calculate_metrics(self.benchmark_name, inference_outputs[task], ground_truth, self.benchmark_cfgs.answer_extractor, self.benchmark_cfgs.metrics)
-
-        if len(task_list) > 1:
-            overall_result = self.calculate_overall_metrics(self.benchmark_cfgs.overall_metrics, result=evaluation_results)
-        else:
-            overall_result = {}
-        # evaluation_results['Overall'] = overall_result
-        self.display_benchmark_results(self.benchmark_name, evaluation_results)
-        if overall_result != {}:
-            self.display_benchmark_results(self.benchmark_name, overall_result)
-        self.save_benchmark_details(self.output_path, self.benchmark_name, input_data, evaluation_details)
-        return evaluation_details, evaluation_results, overall_result
+        for task, inference_input in input_data.items():
+            for item in inference_input:
+                item.ref_answer = RegexMatchNumber(additional_pattern=r"{original_pattern}", match_index=-1).apply([item.ref_answer])[0]
+            
+        return input_data
